Fixed double Login Page
